Renamed bp_classes to Spree::Brightpearl namespace
