+++ app/models/spree/payment_decorator.rb
@@ -3,7 +3,7 @@ module Spree
     Spree::Payment.state_machine.after_transition :to => :completed, do: :save_to_brightpearl
 
     def save_to_brightpearl
-      BpPayment.delay.create(id) if order.brightpearl_id
+      Brightpearl::Payment.delay.create(id) if order.brightpearl_id
     end
   end
 end
