+++ spec/models/spree/brightpearl/contact_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe Spree::BpContact do
+describe Spree::Brightpearl::Contact do
   let(:user) { build :user }
 
   subject { described_class.new(user, 100) }
