+++ spec/models/spree/order_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe Spree::BpOrder do
+describe Spree::Order do
   before do
     Spree::Config[:brightpearl_email] = 'steven@crowdint.com'
     Spree::Config[:brightpearl_id] = 'crowdint'
@@ -9,11 +9,13 @@ describe Spree::BpOrder do
 
   let(:order) { VCR.use_cassette('bp/authorise') { create(:order_ready_to_complete) }}
 
-  it 'queue a new job' do
-    BpOrderWorker.should_receive(:perform_async).with(order.id)
-    
-    VCR.use_cassette('bp/order_create') do
-      order.next!
+  describe '#create_brightpearl_order' do
+    it 'queue a new job' do
+      BpOrderWorker.should_receive(:perform_async).with(order.id)
+
+      VCR.use_cassette('bp/order_create') do
+        order.next!
+      end
     end
   end
 end
